💗🍏 Checkpoint ./commands/Minecraft/start.js:21789943/243 ./watch.json:21789943/2

+++ commands/Minecraft/start.js
@@ -37,7 +37,7 @@ msg.channel.send(`**${msg.author.tag}**, Your nickname has been stored in the da
   emb1.delete()
   const lkk = await msg.channel.send(jk)
   try {
-  await msg.channel.awaitMessages(filter, { max: 1, time: 20000, errors:['time']}).then(res => {
+  await msg.channel.awaitMessages(filter, { max: 1, time: 20000, errors:['time']}).then((res) => {
     
     msg.delete()
     
@@ -52,11 +52,11 @@ msg.channel.send(`**${msg.author.tag}**, Your nickname has been stored in the da
     } else if(res.first().content.toLowerCase() === "female") {
       data.push("female")
      return msg.channel.send(`**${msg.author.tag}**, Your gender has been stored in the database! (\`Female\`)`).then(m => m.delete({timeout:5000}))
-    } else {
+    } else if(res.first.content.toLowerCase === msg.content.split(" ")){  
       lkk.delete()
     return msg.channel.send(`**${msg.author.tag}**, Gender is not valid!`).then(w => w.delete({timeout:5000}))
-     
     }
+    
             
   })
 } catch(err) {
